+++ src/Pdf.php
@@ -9,20 +9,36 @@ use mikehaertl\tmp\File;
  * This class is a slim wrapper around wkhtmltopdf.
  *
  * @author Michael Härtl <haertl.mike@gmail.com>
- * @version 2.0.2-dev
  * @license http://www.opensource.org/licenses/MIT
  */
 class Pdf
 {
-    // Regular expression to detect HTML strings
-    const REGEX_HTML = '/<html/i';
+    // Type hints for `addPage()` and `addCover()`
+    const TYPE_HTML = 'html';
+    const TYPE_XML = 'xml';
 
-    // prefix for tmp files
+    // Regular expression to detect HTML strings
+    const REGEX_HTML = '/<(?:!doctype )?html/i';
+
+    // Regular expression to detect XML strings
+    const REGEX_XML = '/<\??xml/i';
+
+    // Regular expression to detect URL strings
+    const REGEX_URL = '/^(https?:)?\/\//i';
+
+    // Regular expression to detect options that expect an URL or a file name,
+    // so we need to create a tmp file for the content.
+    const REGEX_OPTS_TMPFILE = '/^((header|footer)-html|(xsl|user)-style-sheet)$/i';
+
+    // Prefix for tmp files
     const TMP_PREFIX = 'tmp_wkhtmlto_pdf_';
 
+    // Maximum length of a file path if PHP_MAXPATHLEN is not defined
+    const MAX_PATHLEN = 255;
+
     /**
-     * @var string the name of the `wkhtmltopdf` binary. Default is `wkhtmltopdf`. You can also
-     * configure a full path here.
+     * @var string the name of the `wkhtmltopdf` binary. Default is
+     * `wkhtmltopdf`. You can also configure a full path here.
      */
     public $binary = 'wkhtmltopdf';
 
@@ -32,17 +48,20 @@ class Pdf
     public $commandOptions = array();
 
     /**
-     * @var string|null the directory to use for temporary files. If null (default) the dir is autodetected.
+     * @var string|null the directory to use for temporary files. If null
+     * (default) the dir is autodetected.
      */
     public $tmpDir;
 
     /**
-     * @var bool whether to ignore any errors if some PDF file was still created. Default is false.
+     * @var bool whether to ignore any errors if some PDF file was still
+     * created. Default is false.
      */
     public $ignoreWarnings = false;
 
     /**
-     * @var bool whether the old version 9 of wkhtmltopdf is used (slightly different syntax). Default is false.
+     * @var bool whether the old version 9 of wkhtmltopdf is used (slightly
+     * different syntax). Default is false.
      */
     public $version9 = false;
 
@@ -52,7 +71,8 @@ class Pdf
     protected $_isCreated = false;
 
     /**
-     * @var array global options for wkhtmltopdf as array('--opt1', '--opt2'=>'val', ...)
+     * @var array global options for `wkhtmltopdf` as `['--opt1', '--opt2' =>
+     * 'val', ...]`
      */
     protected $_options = array();
 
@@ -62,13 +82,14 @@ class Pdf
     protected $_objects = array();
 
     /**
-     * @var mikehaertl\tmp\File the temporary PDF file
+     * @var \mikehaertl\tmp\File the temporary PDF file
      */
     protected $_tmpPdfFile;
 
     /**
-     * @var array list of tmp file objects. This is here to keep a reference to File and thus avoid
-     * too early call of File::__destruct() if the file is not referenced anymore.
+     * @var \mikehaertl\tmp\File[] list of tmp file objects. This is here to
+     * keep a reference to `File` and thus avoid too early call of
+     * [[File::__destruct]] if the file is not referenced anymore.
      */
     protected $_tmpFiles = array();
 
@@ -83,9 +104,10 @@ class Pdf
     protected $_error = '';
 
     /**
-     * @param array|string $options global options for wkhtmltopdf or page URL, HTML or PDF/HTML filename
+     * @param array|string $options global options for wkhtmltopdf, a page URL,
+     * a HTML string or a filename
      */
-    public function __construct($options=null)
+    public function __construct($options = null)
     {
         if (is_array($options)) {
             $this->setOptions($options);
@@ -97,30 +119,35 @@ class Pdf
     /**
      * Add a page object to the output
      *
-     * @param string $input either a URL, a HTML string or a PDF/HTML filename
+     * @param string $input either a URL, a HTML string or a filename
      * @param array $options optional options for this page
-     * @return Pdf the Pdf instance for method chaining
+     * @param string|null $type a type hint if the input is a string of known
+     * type. This can either be `TYPE_HTML` or `TYPE_XML`. If `null` (default)
+     * the type is auto detected from the string content.
+     * @return static the Pdf instance for method chaining
      */
-    public function addPage($input,$options=array())
+    public function addPage($input, $options = array(), $type = null)
     {
-        $options = $this->processOptions($options);
-        $options['inputArg'] = $this->processInput($input);
-        $this->_objects[] = $options;
+        $options['inputArg'] = $this->ensureUrlOrFile($input, $type);
+        $this->_objects[] = $this->ensureUrlOrFileOptions($options);
         return $this;
     }
 
     /**
      * Add a cover page object to the output
      *
-     * @param string $input either a URL, a HTML string or a PDF/HTML filename
+     * @param string $input either a URL, a HTML string or a filename
      * @param array $options optional options for the cover page
-     * @return Pdf the Pdf instance for method chaining
+     * @param string|null $type a type hint if the input is a string of known
+     * type. This can either be `TYPE_HTML` or `TYPE_XML`. If `null` (default)
+     * the type is auto detected from the string content.
+     * @return static the Pdf instance for method chaining
      */
-    public function addCover($input,$options=array())
+    public function addCover($input, $options = array(), $type = null)
     {
-        $options['input'] = ($this->version9 ? '--' : '').'cover';
-        $options['inputArg'] = $this->processInput($input);
-        $this->_objects[] = $options;
+        $options['input'] = ($this->version9 ? '--' : '') . 'cover';
+        $options['inputArg'] = $this->ensureUrlOrFile($input, $type);
+        $this->_objects[] = $this->ensureUrlOrFileOptions($options);
         return $this;
     }
 
@@ -128,12 +155,12 @@ class Pdf
      * Add a TOC object to the output
      *
      * @param array $options optional options for the table of contents
-     * @return Pdf the Pdf instance for method chaining
+     * @return static the Pdf instance for method chaining
      */
-    public function addToc($options=array())
+    public function addToc($options = array())
     {
-        $options['input'] = ($this->version9 ? '--' : '')."toc";
-        $this->_objects[] = $options;
+        $options['input'] = ($this->version9 ? '--' : '') . 'toc';
+        $this->_objects[] = $this->ensureUrlOrFileOptions($options);
         return $this;
     }
 
@@ -158,32 +185,59 @@ class Pdf
     /**
      * Send PDF to client, either inline or as download (triggers PDF creation)
      *
-     * @param string|null $filename the filename to send. If empty, the PDF is streamed inline.
-     * @param bool $inline whether to force inline display of the PDF, even if filename is present.
+     * @param string|null $filename the filename to send. If empty, the PDF is
+     * streamed inline.
+     * @param bool $inline whether to force inline display of the PDF, even if
+     * filename is present.
+     * @param array $headers a list of additional HTTP headers to send in the
+     * response as an array. The array keys are the header names like
+     * 'Cache-Control' and the array values the header value strings to send.
+     * Each array value can also be another array of strings if the same header
+     * should be sent multiple times. This can also be used to override
+     * automatically created headers like 'Expires' or 'Content-Length'. To suppress
+     * automatically created headers, `false` can also be used as header value.
      * @return bool whether PDF was created successfully
      */
-    public function send($filename=null,$inline=false)
+    public function send($filename = null, $inline = false, $headers = array())
     {
         if (!$this->_isCreated && !$this->createPdf()) {
             return false;
         }
-        $this->_tmpPdfFile->send($filename, 'application/pdf', $inline);
+        $this->_tmpPdfFile->send($filename, 'application/pdf', $inline, $headers);
         return true;
+    }
+
+    /**
+     * Get the raw PDF contents (triggers PDF creation).
+     * @return string|bool The PDF content as a string or `false` if the PDF
+     * wasn't created successfully.
+     */
+    public function toString()
+    {
+        if (!$this->_isCreated && !$this->createPdf()) {
+            return false;
+        }
+        return file_get_contents($this->_tmpPdfFile->getFileName());
     }
 
     /**
      * Set global option(s)
      *
      * @param array $options list of global PDF options to set as name/value pairs
-     * @return Pdf the Pdf instance for method chaining
+     * @return static the Pdf instance for method chaining
      */
-    public function setOptions($options=array())
+    public function setOptions($options = array())
     {
-        $options = $this->processOptions($options);
-        foreach ($options as $key=>$val) {
+        // #264 tmpDir must be set before calling ensureUrlOrFileOptions
+        if (isset($options['tmpDir'])) {
+            $this->tmpDir = $options['tmpDir'];
+            unset($options['tmpDir']);
+        }
+        $options = $this->ensureUrlOrFileOptions($options);
+        foreach ($options as $key => $val) {
             if (is_int($key)) {
                 $this->_options[] = $val;
-            } elseif ($key[0]!=='_' && property_exists($this, $key)) {
+            } elseif ($key[0] !== '_' && property_exists($this, $key)) {
                 $this->$key = $val;
             } else {
                 $this->_options[$key] = $val;
@@ -197,7 +251,7 @@ class Pdf
      */
     public function getCommand()
     {
-        if ($this->_command===null) {
+        if ($this->_command === null) {
             $options = $this->commandOptions;
             if (!isset($options['command'])) {
                 $options['command'] = $this->binary;
@@ -220,7 +274,7 @@ class Pdf
      */
     public function getPdfFilename()
     {
-        if ($this->_tmpPdfFile===null) {
+        if ($this->_tmpPdfFile === null) {
             $this->_tmpPdfFile = new File('', '.pdf', self::TMP_PREFIX, $this->tmpDir);
         }
         return $this->_tmpPdfFile->getFileName();
@@ -246,7 +300,7 @@ class Pdf
         $command->addArg($fileName, null, true);    // Always escape filename
         if (!$command->execute()) {
             $this->_error = $command->getError();
-            if (!(file_exists($fileName) && filesize($fileName)!==0 && $this->ignoreWarnings)) {
+            if (!(file_exists($fileName) && filesize($fileName) !== 0 && $this->ignoreWarnings)) {
                 return false;
             }
         }
@@ -255,28 +309,59 @@ class Pdf
     }
 
     /**
-     * @param string $input
-     * @return mikehaertl\tmp\File|string a File object if the input is a html string. The unchanged input otherwhise.
+     * This method creates a temporary file if the passed argument is neither a
+     * File instance or URL nor contains XML or HTML and is also not a valid
+     * file name.
+     *
+     * @param string|File $input the input argument File to check
+     * @param string|null $type a type hint if the input is a string of known
+     * type. This can either be `TYPE_HTML` or `TYPE_XML`. If `null` (default)
+     * the type is auto detected from the string content.
+     * @return \mikehaertl\tmp\File|string a File object if the input is a HTML
+     * or XML string. The unchanged input otherwhise.
      */
-    protected function processInput($input)
+    protected function ensureUrlOrFile($input, $type = null)
     {
-        if (preg_match(self::REGEX_HTML, $input)) {
-            return $this->_tmpFiles[] = new File($input, '.html', self::TMP_PREFIX, $this->tmpDir);
-        } else {
+        if ($input instanceof File) {
+            $this->_tmpFiles[] = $input;
             return $input;
+        } elseif (preg_match(self::REGEX_URL, $input)) {
+            return $input;
+        } elseif ($type === self::TYPE_XML || $type === null && preg_match(self::REGEX_XML, $input)) {
+            $ext = '.xml';
+        } else {
+            // First check for obvious HTML content to avoid is_file() as much
+            // as possible as it can trigger open_basedir restriction warnings
+            // with long strings.
+            $isHtml = $type === self::TYPE_HTML || preg_match(self::REGEX_HTML, $input);
+            if (!$isHtml) {
+                $maxPathLen = defined('PHP_MAXPATHLEN') ?
+                    constant('PHP_MAXPATHLEN') : self::MAX_PATHLEN;
+                if (strlen($input) <= $maxPathLen && is_file($input)) {
+                    return $input;
+                }
+            }
+            $ext = '.html';
         }
+        $file = new File($input, $ext, self::TMP_PREFIX, $this->tmpDir);
+        $this->_tmpFiles[] = $file;
+        return $file;
     }
 
     /**
      * @param array $options list of options as name/value pairs
-     * @return array options with raw content converted to tmp files where neccessary
+     * @return array options with raw HTML/XML/String content converted to tmp
+     * files where neccessary
      */
-    protected function processOptions($options=array())
+    protected function ensureUrlOrFileOptions($options = array())
     {
-        foreach ($options as $key=>$val) {
-            $urlRequired = preg_match('/^(header|footer)-html$/', $key);
-            if ($urlRequired && !(is_file($val) || preg_match('/^(https?:)?\/\//i',$val) || $val===strip_tags($val))) {
-                $options[$key] = new File($val, '.html', self::TMP_PREFIX, $this->tmpDir);
+        foreach ($options as $key => $val) {
+            // Some options expect a URL or a file name, so check if we need a temp file
+            if (is_string($val) && preg_match(self::REGEX_OPTS_TMPFILE, $key) ) {
+                $file = $this->ensureUrlOrFile($val);
+                if ($file instanceof File) {
+                    $options[$key] = $file;
+                }
             }
         }
         return $options;